Bluetooth: Mesh: Filter duplicates in brg subnets list
This commit adds functionality to filter out duplicate entries in the
Bridged Subnets List message.

Because the server is required to be able to index into the filtered
list, we need to keep track of all unique pairs we have seen so far
while constructing the message, even those that show up at indexes
earlier than the requested start index.

This is handled by creating a long buffer able to contain a full,
potentially unfiltered list, and then altering the buffer to point to a
sub-range of this data to extract a message containing the relevant part
of the list indexed by the requested start index before sending.

Signed-off-by: Ludvig Jordet <ludvig.jordet@nordicsemi.no>